PR comments: move formula functions to molutil
