+++ qcelemental/molutil/__init__.py
@@ -1,2 +1,3 @@
 from .align import B787, compute_scramble, kabsch_align
 from .connectivity import guess_connectivity
+from .molecular_formula import order_molecular_formula, molecular_formula_from_symbols
